+++ tsp/src/ils/mod.rs
@@ -8,9 +8,9 @@ use instance_reader::Instance;
 use local_search::local_search;
 use perturbation::perturbation;
 
-pub fn ils(max_iter: usize, max_iter_ils: usize, instance: &Instance) -> Solution {
+pub fn ils(max_iter: u32, max_iter_ils: u32, instance: &Instance) -> Solution {
     let mut best = Solution {
-        value: usize::MAX,
+        value: u32::MAX,
         ..Default::default()
     };
