+++ tsp/src/solution.rs
@@ -3,7 +3,7 @@ use instance_reader::Instance;
 #[derive(Debug, Clone, Default)]
 pub struct Solution {
     pub sequence: Vec<usize>,
-    pub value: usize,
+    pub value: u32,
 }
 
 impl Solution {
